MSys patch for DGD
I've tried sending this once before, but apparently the mail server
doesn't like attachments, so I'm inlining it this time.

Anyways, I've recently finished a patch to compile DGD on the MSys
system (I didn't like how 'heavy' MSVC++ was). I've tested it
(compiled, executed, telnetted in, issued the shutdown command), and
the patch seems to work on my system. However, I do not know how the
patch will act on a *nix system (I expect it to be fine, but I might
have goofed somewhere), and I don't know how the patch will act with a
non-GNU make. Also, should I submit this to DGD hub, or make a
sourceforge project to hold the patch, or what? Here's the contents of
the patch file: