DGD For MSys, mark 2
I've redone my patch to compile DGD on the MSys environment. This
version was diff'd against a .zip copy that it looks like I grabbed on
the 27th of December. I didn't grab a current copy of the kernellib,
so this is NOT a tested build. However, I seem to have learned a good
deal from last time, because I only altered 3 files this time, and one
of those was just to insert an obnoxious warning into the README.

I've made certain to add plenty of comments.

Note that like the previous version, I'm only mailing this in case
someone comes looking for such a patch. I haven't tested the resulting
executable, and therefor don't know if I'm done or not. Depending on
the outcome of certain inquiries on the MinGW/MSys mailing list, I'm
likely to modify the host detection code.

I previously tried to send the patch as an attachment, and apparently
it got rejected by the mailer-daemon , so here it is inlined: