Cygwin detection, attempt 1
Revised the comment in front of the initial Windows detection code.
Changed the "MSys required" to a warning, and expanded to mention Cygwin support too. Make should now continue past this on Cygwin.
Added Cygwin detection as an "else ifeq" to the MSys detection. Both set the same Make vars.